move preconditioner method in NystromSolver to separate module
